Move docstrings from implementation to interface
