+++ eth/vm/chain_context.py
@@ -7,10 +7,6 @@ from eth.validation import (
 
 
 class ChainContext(ChainContextAPI):
-    """
-    This immutable object houses chain information that remains constant for the entire context of
-    the VM execution.
-    """
     __slots__ = ['_chain_id']
 
     def __init__(self, chain_id: Optional[int]) -> None:
