+++ newsfragments/1880.doc.rst
@@ -1,0 +1,7 @@
+Add missing API docs for ``MiningChain``.
+Add missing API docs for ``eth.db.*``
+Add missing API docs for ``Constantinople``, ``Petersburg`` and ``Istanbul`` forks
+Move all docstrings that aren't overly specific to a particular implementation from
+the implementation to the interface. This has the effect that the docstring will
+appear both on the interface as well as on the implementation except for when the
+implementation overwrites the docstring with a more specific descriptions.
